Changed the md-tabs into semantic-ui tabs
The reason what that with md-tabs the responsiveness created too much trouble (despite the fact the tabs could be scrolled, the names disappeared and were unreadable). Now I've exchanged that for semantic-ui which not only are more universal (do not need angular as dependency), but also are much more flexible (I could do the responsiveness by treating the tabs as grids). The only drawback is that there is no built-in swiping between tabs, but that can be implemented in the future.

Added animations and icons for semantic-ui tabs
Added responsiveness for semantic-ui tabs

Added icons for the tabs.
Changed the background for the whole application for *slightly* darker white
Added an application.scss file and that's the place you want to put all the things you want to see in the whole application